+++ .gitignore
@@ -1,0 +1,4 @@
+# Files to ignore
+
+sorted-set/__pycache__/
+env
